Fix pipelined-client so it works after a timedout request
When the connection is down initially, (connection) returns a
result-channel with an error. Since the pipeline didn't have a
error-handler this pops up causing receive-in-order to finish. All
subsequent requests will not be processed.

An error-handler is provided so it retries if the result has not been
completed, i.e, the timeout has expired.

Besides an exception was happening when reentering the requests
receive-in-order when a connection is returned but it's drained. The
timeout was nil and this caused an exception that aborted the
receive-in-order.

Instead of looking at timeout to determine if the execution must
continue a new function, has-completed?, is defined. It looks at the
result-channel instead of the timeout value.